♻️ refactor: 컴포넌트 생성을 단일 useCreateAndMountApp 훅으로 통합
- CartUpdateService 클래스 제거, useCartUpdater 함수로 통합
- 함수형 접근 방식으로 카트 로직 구조 개선
- 기존 API (calculateCart, updatePricesAndCalculate) 유지
- JSDoc 주석 추가로 명확성 향상
- safeUpdateCart 메서드 제거로 인터페이스 간소화
